clean up login component
